# swagger.yaml
swagger: '2.0'
info:
  description: StatusTick API documentation
  version: 1.0.0
  title: StatusTick API
  contact:
    email: mustafa@statustick.com

  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

# tags are used for organizing operations
...
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
paths:
...
  /auth/account:
    get:
      tags:
      - auth
      summary: Returns authorised account
      operationId: authAccount
      description: Authorised account
      produces:
      - application/json
      security:
        - Bearer: []
      responses:
        200:
          description: Successfully authorized user
          schema:
            $ref: '#/definitions/AccountResponse'
...
definitions:
...
  AccountResponse:
    type: object
    properties:
      success:
        type: boolean
        example: true
      account:
        $ref: '#/definitions/Account'
...
  User:
    properties:
      id:
        type: string
        example: 12345678abc87654321
      username:
        type: string
        example:  myusername
      createdAt:
        type: string
        format: ISO8601
        example: '2018-12-31T23:59:59.000Z'
  Account:
    properties:
      id:
        type: string
        example: 12345678abc87654321
      name:
        type: string
        example: 'Account Name'
      createdAt:
        type: string
        format: ISO8601
        example: '2018-12-31T23:59:59.000Z'
      updatedAt:
        type: string
        format: ISO8601
        example: '2018-12-31T23:59:59.000Z'
      user:
        $ref: '#/definitions/User'
      users:
        type: array
        items:
          $ref: '#/definitions/AccountRole'
      timezone:
        type: string
        example: 'Europe/Istanbul'
      token:
        type: string
        example: abc.abcabc.abc
  AccountRole:
    properties:
      permission:
        type: string
        example: Admin
      user:
        $ref: '#/definitions/User'
      pending:
        type: boolean
        example: false
...
schemes:
 - https
host: api.statustick.com